+++ tests/plugins/piccolo_orm/test_piccolo_orm_plugin_openapi_specs.py
@@ -1,5 +1,5 @@
 from starlite import Starlite
-from starlite.plugins.piccolo_orm import PiccoloORMPlugin
+from starlite.contrib.piccolo_orm import PiccoloORMPlugin
 from tests.plugins.piccolo_orm.endpoints import (
     create_concert,
     retrieve_studio,
